backend/requirements.txt
fastapi
uvicorn[standard]
httpx
elasticsearch<9.0.0,>=8.0.0
pymongo
python-dotenv
sentence-transformers
torch
huggingface_hub[hf_xet]
pandas
numpy
openai
python-docx
motor
python-multipart
langchain
langchain-community
langchain-openai
requests
pypdf
anthropic>=0.21.3
pdfplumber
pytesseract
pdf2image
pillow
sseclient-py
sse-starlette>=1.0.6
pdfminer.six
PyPDF2